readyup|feat!: Move check-utils to dedicated subpath export
Check utilities no longer re-export from the `readyup` package root and now live exclusively at the new `readyup/check-utils` subpath. Kit authors must split imports: authoring helpers stay on `readyup`; check utilities like `fileExists` and `discoverWorkspaces` move to `readyup/check-utils`.

`readyup/check-utils` is the stable, versioned surface for kit-author imports of check utilities. Compiled kits embed nothing of readyup itself; the `rdy` runner satisfies `readyup` and `readyup/*` imports at runtime via its module-resolution hook. Kits are therefore version-coupled to the runner across breaking boundaries: after upgrading readyup across a major, recompile kits with `rdy compile` so any newly-shipped or changed check utilities are picked up.

🚨 **Breaking:** check utilities are no longer importable from `readyup` — update kit imports to `readyup/check-utils`.
