+++ .readyup/manifest.json
@@ -5,7 +5,7 @@
       "name": "demo",
       "path": "kits/demo.js",
       "source": "kits/demo.ts",
-      "targetHash": "6ac8fdc4"
+      "targetHash": "ec1235a9"
     }
   ]
 }
