+++ packages/readyup/src/compile/compileConfig.ts
@@ -35,12 +35,6 @@ function deriveOutputPath(inputPath: string): string {
  * are resolved at runtime by the `rdy` runner's module-resolution hook
  * (`readyupResolverHook.ts`), which routes them to the runner's own readyup
  * installation. Prepends a generated-file header comment to the output.
- *
- * Note: `readyup/<subpath>` specifiers are listed external so that the bundler
- * leaves them as live imports, and the hook routes them correctly at runtime.
- * However, the `readyup` package's `exports` map only declares `"."` at present;
- * subpath exports (e.g., `readyup/check-utils`) ship in PR #85. Until then,
- * compiled kits should import from `readyup` only.
  */
 export async function compileConfig(inputPath: string, outputPath?: string): Promise<CompileResult> {
   const resolvedInput = path.resolve(inputPath);
